ci: lint tests and benches with clippy (--all-targets) and fix the fallout
CI ran `cargo clippy --workspace` only over lib/bin targets, so test,
bench and example code was never linted. `--all-targets` reported ~4.5k
errors. Fixes:

- clippy.toml: `allow-unwrap-in-tests = true`. `unwrap_used` stays denied
  in shipped code; in tests a panic on Err/None is the failure report.
  This covered ~4.1k of the errors, all of the oasis-js ones included.
- field_reassign_with_default (~150 test setups): rewritten to
  struct-update syntax (`T { a, ..Default::default() }`).
- Tautological / constant assertions: the blur filter test's
  `x < y || x >= y` checks now assert the colour really desaturates;
  checks on constants became `const { assert!(..) }`.
- tv_controller tests used `read()` where a short read would pass
  silently; they now use `read_exact()`.
- Machine-applicable lints via `clippy --fix` (while-let loops,
  collapsible ifs, manual range/div_ceil, test modules moved to file end,
  etc.). `module_inception` in two `tests.rs` files: flattened the inner
  `mod tests`, and gated active_theme's `mod tests;` behind cfg(test).
- Targeted allows, each with a reason: benches (`unwrap_used` in setup),
  the oasis-ffi test module (one SAFETY note for its ~130 unsafe calls
  into the crate's own exports, plus the null checks it relies on),
  `clone_on_copy` in the test that exists to exercise Clone, and
  `ptr_arg` where `ListView<String>`'s callback type forces `&String`.

CI (main + PR), the nightly advisory step, the pre-commit hooks,
run-ci.sh and the docs now all use
`cargo clippy --workspace --all-targets -- -D warnings`.

+++ crates/oasis-backend-sdl/src/lib.rs
@@ -1871,7 +1871,7 @@ mod tests {
                 if ttf.has_glyph(ch) {
                     ttf.advance(ch, 16.0).max(0) as u32
                 } else {
-                    oasis_types::bitmap_font::glyph_advance_scaled(ch, 16) as u32
+                    oasis_types::bitmap_font::glyph_advance_scaled(ch, 16)
                 }
             })
             .sum();
